Lab_7/app/requirements.txt
Faker==30.8.1
prettytable==3.11.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
six==1.16.0
typing_extensions==4.12.2
wcwidth==0.2.13
